[Autocomplete] Move useAutocomplete to unstyled package
