+++ packages/material-ui-unstyled/src/AutocompleteUnstyled/index.d.ts
@@ -1,0 +1,2 @@
+export { default as useAutocomplete } from './useAutocomplete';
+export * from './useAutocomplete';
