+++ packages/material-ui-unstyled/src/AutocompleteUnstyled/index.js
@@ -1,0 +1,1 @@
+export { default as useAutocomplete, createFilterOptions } from './useAutocomplete';
