+++ packages/material-ui-unstyled/src/AutocompleteUnstyled/useAutocomplete.spec.ts
@@ -1,4 +1,4 @@
-import useAutocomplete, { FilterOptionsState } from '@material-ui/core/useAutocomplete';
+import { useAutocomplete, FilterOptionsState } from '@material-ui/unstyled/AutocompleteUnstyled';
 import { expectType } from '@material-ui/types';
 
 interface Person {
