+++ packages/material-ui-unstyled/src/AutocompleteUnstyled/useAutocomplete.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { createClientRender, screen, ErrorBoundary } from 'test/utils';
-import useAutocomplete, { createFilterOptions } from '@material-ui/core/useAutocomplete';
+import { useAutocomplete, createFilterOptions } from '@material-ui/unstyled/AutocompleteUnstyled';
 
 describe('useAutocomplete', () => {
   const render = createClientRender();
